+++ squeal-postgresql/src/Squeal/PostgreSQL/PQ/Oid.hs
@@ -14,6 +14,7 @@ statement parameters.
     AllowAmbiguousTypes
   , DataKinds
   , FlexibleInstances
+  , MultiParamTypeClasses
   , PolyKinds
   , ScopedTypeVariables
   , TypeApplications
@@ -25,13 +26,17 @@ module Squeal.PostgreSQL.PQ.Oid
     LibPQ.Oid
   , OidOf (..)
   , OidOfArray (..)
+  , oidOfTypedef
+  , oidOfArrayTypedef
   , OidOfNull (..)
   , OidOfField (..)
   ) where
 
+import Control.Monad.Reader
 import GHC.TypeLits
 
 import qualified Database.PostgreSQL.LibPQ as LibPQ
+import qualified Generics.SOP as SOP
 
 import Squeal.PostgreSQL.Alias
 import Squeal.PostgreSQL.Schema
@@ -41,91 +46,99 @@ import Squeal.PostgreSQL.Schema
 -- >>> :set -XTypeApplications
 -- >>> oidOf @'PGbool
 -- Oid 16
-class OidOf (pg :: PGType) where
-  oidOf :: LibPQ.Oid
+class OidOf (db :: SchemasType) (pg :: PGType) where
+  oidOf :: ReaderT (SOP.K LibPQ.Connection db) IO LibPQ.Oid
 -- | The `LibPQ.Oid` of an array
-class OidOfArray (pg :: PGType) where
-  oidOfArray :: LibPQ.Oid
-instance OidOfArray pg => OidOf ('PGvararray (null pg)) where
-  oidOf = oidOfArray @pg
-instance OidOfArray pg => OidOf ('PGfixarray dims (null pg)) where
-  oidOf = oidOfArray @pg
+class OidOfArray (db :: SchemasType) (pg :: PGType) where
+  oidOfArray :: ReaderT (SOP.K LibPQ.Connection db) IO LibPQ.Oid
+instance OidOfArray db pg => OidOf db ('PGvararray (null pg)) where
+  oidOf = oidOfArray @db @pg
+instance OidOfArray db pg => OidOf db ('PGfixarray dims (null pg)) where
+  oidOf = oidOfArray @db @pg
 -- | The `LibPQ.Oid` of a `NullType`
-class OidOfNull (ty :: NullType) where
-  oidOfNull :: LibPQ.Oid
-instance OidOf pg => OidOfNull (null pg) where
-  oidOfNull = oidOf @pg
+class OidOfNull (db :: SchemasType) (ty :: NullType) where
+  oidOfNull :: ReaderT (SOP.K LibPQ.Connection db) IO LibPQ.Oid
+instance OidOf db pg => OidOfNull db (null pg) where
+  oidOfNull = oidOf @db @pg
 -- | The `LibPQ.Oid` of a field
-class OidOfField (field :: (Symbol, NullType)) where
-  oidOfField :: LibPQ.Oid
-instance OidOfNull ty => OidOfField (fld ::: ty) where
-  oidOfField = oidOfNull @ty
+class OidOfField (db :: SchemasType) (field :: (Symbol, NullType)) where
+  oidOfField :: ReaderT (SOP.K LibPQ.Connection db) IO LibPQ.Oid
+instance OidOfNull db ty => OidOfField db (fld ::: ty) where
+  oidOfField = oidOfNull @db @ty
 
-instance OidOf 'PGbool where oidOf = LibPQ.Oid 16
-instance OidOfArray 'PGbool where oidOfArray = LibPQ.Oid 1000
-instance OidOf 'PGint2 where oidOf = LibPQ.Oid 21
-instance OidOfArray 'PGint2 where oidOfArray = LibPQ.Oid 1005
-instance OidOf 'PGint4 where oidOf = LibPQ.Oid 23
-instance OidOfArray 'PGint4 where oidOfArray = LibPQ.Oid 1007
-instance OidOf 'PGint8 where oidOf = LibPQ.Oid 20
-instance OidOfArray 'PGint8 where oidOfArray = LibPQ.Oid 1016
-instance OidOf 'PGnumeric where oidOf = LibPQ.Oid 1700
-instance OidOfArray 'PGnumeric where oidOfArray = LibPQ.Oid 1231
-instance OidOf 'PGfloat4 where oidOf = LibPQ.Oid 700
-instance OidOfArray 'PGfloat4 where oidOfArray = LibPQ.Oid 1021
-instance OidOf 'PGfloat8 where oidOf = LibPQ.Oid 701
-instance OidOfArray 'PGfloat8 where oidOfArray = LibPQ.Oid 1022
-instance OidOf 'PGmoney where oidOf = LibPQ.Oid 790
-instance OidOfArray 'PGmoney where oidOfArray = LibPQ.Oid 791
-instance OidOf ('PGchar n) where oidOf = LibPQ.Oid 18
-instance OidOfArray ('PGchar n) where oidOfArray = LibPQ.Oid 1002
-instance OidOf ('PGvarchar n) where oidOf = LibPQ.Oid 1043
-instance OidOfArray ('PGvarchar n) where oidOfArray = LibPQ.Oid 1015
-instance OidOf 'PGtext where oidOf = LibPQ.Oid 25
-instance OidOfArray 'PGtext where oidOfArray = LibPQ.Oid 1009
-instance OidOf 'PGbytea where oidOf = LibPQ.Oid 17
-instance OidOfArray 'PGbytea where oidOfArray = LibPQ.Oid 1001
-instance OidOf 'PGtimestamp where oidOf = LibPQ.Oid 1114
-instance OidOfArray 'PGtimestamp where oidOfArray = LibPQ.Oid 1115
-instance OidOf 'PGtimestamptz where oidOf = LibPQ.Oid 1184
-instance OidOfArray 'PGtimestamptz where oidOfArray = LibPQ.Oid 1185
-instance OidOf 'PGdate where oidOf = LibPQ.Oid 1082
-instance OidOfArray 'PGdate where oidOfArray = LibPQ.Oid 1182
-instance OidOf 'PGtime where oidOf = LibPQ.Oid 1083
-instance OidOfArray 'PGtime where oidOfArray = LibPQ.Oid 1183
-instance OidOf 'PGtimetz where oidOf = LibPQ.Oid 1266
-instance OidOfArray 'PGtimetz where oidOfArray = LibPQ.Oid 1270
-instance OidOf 'PGinterval where oidOf = LibPQ.Oid 1186
-instance OidOfArray 'PGinterval where oidOfArray = LibPQ.Oid 1187
-instance OidOf 'PGuuid where oidOf = LibPQ.Oid 2950
-instance OidOfArray 'PGuuid where oidOfArray = LibPQ.Oid 2951
-instance OidOf 'PGinet where oidOf = LibPQ.Oid 869
-instance OidOfArray 'PGinet where oidOfArray = LibPQ.Oid 1041
-instance OidOf 'PGjson where oidOf = LibPQ.Oid 114
-instance OidOfArray 'PGjson where oidOfArray = LibPQ.Oid 199
-instance OidOf 'PGjsonb where oidOf = LibPQ.Oid 3802
-instance OidOfArray 'PGjsonb where oidOfArray = LibPQ.Oid 3807
-instance OidOf 'PGtsvector where oidOf = LibPQ.Oid 3614
-instance OidOfArray 'PGtsvector where oidOfArray = LibPQ.Oid 3643
-instance OidOf 'PGtsquery where oidOf = LibPQ.Oid 3615
-instance OidOfArray 'PGtsquery where oidOfArray = LibPQ.Oid 3645
-instance OidOf 'PGoid where oidOf = LibPQ.Oid 26
-instance OidOfArray 'PGoid where oidOfArray = LibPQ.Oid 1028
-instance OidOf ('PGrange 'PGint4) where oidOf = LibPQ.Oid 3904
-instance OidOfArray ('PGrange 'PGint4) where oidOfArray = LibPQ.Oid 3905
-instance OidOf ('PGrange 'PGint8) where oidOf = LibPQ.Oid 3926
-instance OidOfArray ('PGrange 'PGint8) where oidOfArray = LibPQ.Oid 3927
-instance OidOf ('PGrange 'PGnumeric) where oidOf = LibPQ.Oid 3906
-instance OidOfArray ('PGrange 'PGnumeric) where oidOfArray = LibPQ.Oid 3907
-instance OidOf ('PGrange 'PGtimestamp) where oidOf = LibPQ.Oid 3908
-instance OidOfArray ('PGrange 'PGtimestamp) where oidOfArray = LibPQ.Oid 3909
-instance OidOf ('PGrange 'PGtimestamptz) where oidOf = LibPQ.Oid 3910
-instance OidOfArray ('PGrange 'PGtimestamptz) where oidOfArray = LibPQ.Oid 3911
-instance OidOf ('PGrange 'PGdate) where oidOf = LibPQ.Oid 3912
-instance OidOfArray ('PGrange 'PGdate) where oidOfArray = LibPQ.Oid 3913
--- instance {-# OVERLAPPABLE #-} OidOf ('PGrange ty) where oidOf = LibPQ.invalidOid
--- instance {-# OVERLAPPABLE #-} OidOfArray ('PGrange ty) where oidOfArray = LibPQ.invalidOid
-instance {-# OVERLAPPABLE #-} OidOf ('PGcomposite row) where oidOf = LibPQ.invalidOid
-instance {-# OVERLAPPABLE #-} OidOfArray ('PGcomposite row) where oidOfArray = LibPQ.invalidOid
--- instance {-# OVERLAPPABLE #-} OidOf ('PGenum labels) where oidOf = LibPQ.invalidOid
--- instance {-# OVERLAPPABLE #-} OidOfArray ('PGenum labels) where oidOfArray = LibPQ.invalidOid
+instance OidOf db 'PGbool where oidOf = pure $ LibPQ.Oid 16
+instance OidOfArray db 'PGbool where oidOfArray = pure $ LibPQ.Oid 1000
+instance OidOf db 'PGint2 where oidOf = pure $ LibPQ.Oid 21
+instance OidOfArray db 'PGint2 where oidOfArray = pure $ LibPQ.Oid 1005
+instance OidOf db 'PGint4 where oidOf = pure $ LibPQ.Oid 23
+instance OidOfArray db 'PGint4 where oidOfArray = pure $ LibPQ.Oid 1007
+instance OidOf db 'PGint8 where oidOf = pure $ LibPQ.Oid 20
+instance OidOfArray db 'PGint8 where oidOfArray = pure $ LibPQ.Oid 1016
+instance OidOf db 'PGnumeric where oidOf = pure $ LibPQ.Oid 1700
+instance OidOfArray db 'PGnumeric where oidOfArray = pure $ LibPQ.Oid 1231
+instance OidOf db 'PGfloat4 where oidOf = pure $ LibPQ.Oid 700
+instance OidOfArray db 'PGfloat4 where oidOfArray = pure $ LibPQ.Oid 1021
+instance OidOf db 'PGfloat8 where oidOf = pure $ LibPQ.Oid 701
+instance OidOfArray db 'PGfloat8 where oidOfArray = pure $ LibPQ.Oid 1022
+instance OidOf db 'PGmoney where oidOf = pure $ LibPQ.Oid 790
+instance OidOfArray db 'PGmoney where oidOfArray = pure $ LibPQ.Oid 791
+instance OidOf db ('PGchar n) where oidOf = pure $ LibPQ.Oid 18
+instance OidOfArray db ('PGchar n) where oidOfArray = pure $ LibPQ.Oid 1002
+instance OidOf db ('PGvarchar n) where oidOf = pure $ LibPQ.Oid 1043
+instance OidOfArray db ('PGvarchar n) where oidOfArray = pure $ LibPQ.Oid 1015
+instance OidOf db 'PGtext where oidOf = pure $ LibPQ.Oid 25
+instance OidOfArray db 'PGtext where oidOfArray = pure $ LibPQ.Oid 1009
+instance OidOf db 'PGbytea where oidOf = pure $ LibPQ.Oid 17
+instance OidOfArray db 'PGbytea where oidOfArray = pure $ LibPQ.Oid 1001
+instance OidOf db 'PGtimestamp where oidOf = pure $ LibPQ.Oid 1114
+instance OidOfArray db 'PGtimestamp where oidOfArray = pure $ LibPQ.Oid 1115
+instance OidOf db 'PGtimestamptz where oidOf = pure $ LibPQ.Oid 1184
+instance OidOfArray db 'PGtimestamptz where oidOfArray = pure $ LibPQ.Oid 1185
+instance OidOf db 'PGdate where oidOf = pure $ LibPQ.Oid 1082
+instance OidOfArray db 'PGdate where oidOfArray = pure $ LibPQ.Oid 1182
+instance OidOf db 'PGtime where oidOf = pure $ LibPQ.Oid 1083
+instance OidOfArray db 'PGtime where oidOfArray = pure $ LibPQ.Oid 1183
+instance OidOf db 'PGtimetz where oidOf = pure $ LibPQ.Oid 1266
+instance OidOfArray db 'PGtimetz where oidOfArray = pure $ LibPQ.Oid 1270
+instance OidOf db 'PGinterval where oidOf = pure $ LibPQ.Oid 1186
+instance OidOfArray db 'PGinterval where oidOfArray = pure $ LibPQ.Oid 1187
+instance OidOf db 'PGuuid where oidOf = pure $ LibPQ.Oid 2950
+instance OidOfArray db 'PGuuid where oidOfArray = pure $ LibPQ.Oid 2951
+instance OidOf db 'PGinet where oidOf = pure $ LibPQ.Oid 869
+instance OidOfArray db 'PGinet where oidOfArray = pure $ LibPQ.Oid 1041
+instance OidOf db 'PGjson where oidOf = pure $ LibPQ.Oid 114
+instance OidOfArray db 'PGjson where oidOfArray = pure $ LibPQ.Oid 199
+instance OidOf db 'PGjsonb where oidOf = pure $ LibPQ.Oid 3802
+instance OidOfArray db 'PGjsonb where oidOfArray = pure $ LibPQ.Oid 3807
+instance OidOf db 'PGtsvector where oidOf = pure $ LibPQ.Oid 3614
+instance OidOfArray db 'PGtsvector where oidOfArray = pure $ LibPQ.Oid 3643
+instance OidOf db 'PGtsquery where oidOf = pure $ LibPQ.Oid 3615
+instance OidOfArray db 'PGtsquery where oidOfArray = pure $ LibPQ.Oid 3645
+instance OidOf db 'PGoid where oidOf = pure $ LibPQ.Oid 26
+instance OidOfArray db 'PGoid where oidOfArray = pure $ LibPQ.Oid 1028
+instance OidOf db ('PGrange 'PGint4) where oidOf = pure $ LibPQ.Oid 3904
+instance OidOfArray db ('PGrange 'PGint4) where oidOfArray = pure $ LibPQ.Oid 3905
+instance OidOf db ('PGrange 'PGint8) where oidOf = pure $ LibPQ.Oid 3926
+instance OidOfArray db ('PGrange 'PGint8) where oidOfArray = pure $ LibPQ.Oid 3927
+instance OidOf db ('PGrange 'PGnumeric) where oidOf = pure $ LibPQ.Oid 3906
+instance OidOfArray db ('PGrange 'PGnumeric) where oidOfArray = pure $ LibPQ.Oid 3907
+instance OidOf db ('PGrange 'PGtimestamp) where oidOf = pure $ LibPQ.Oid 3908
+instance OidOfArray db ('PGrange 'PGtimestamp) where oidOfArray = pure $ LibPQ.Oid 3909
+instance OidOf db ('PGrange 'PGtimestamptz) where oidOf = pure $ LibPQ.Oid 3910
+instance OidOfArray db ('PGrange 'PGtimestamptz) where oidOfArray = pure $ LibPQ.Oid 3911
+instance OidOf db ('PGrange 'PGdate) where oidOf = pure $ LibPQ.Oid 3912
+instance OidOfArray db ('PGrange 'PGdate) where oidOfArray = pure $ LibPQ.Oid 3913
+-- instance {-# OVERLAPPABLE #-} OidOf db ('PGcomposite row) where oidOf = LibPQ.invalidOid
+-- instance {-# OVERLAPPABLE #-} OidOfArray db ('PGcomposite row) where oidOfArray = LibPQ.invalidOid
+-- instance {-# OVERLAPPABLE #-} OidOf db ('PGenum labels) where oidOf = LibPQ.invalidOid
+-- instance {-# OVERLAPPABLE #-} OidOfArray db ('PGenum labels) where oidOfArray = LibPQ.invalidOid
+oidOfTypedef
+  :: (Has sch db schema, Has ty schema pg)
+  => QualifiedAlias sch ty
+  -> ReaderT LibPQ.Connection IO LibPQ.Oid
+oidOfTypedef = undefined
+oidOfArrayTypedef
+  :: (Has sch db schema, Has ty schema pg)
+  => QualifiedAlias sch ty
+  -> ReaderT LibPQ.Connection IO LibPQ.Oid
+oidOfArrayTypedef = undefined
